ReleasableCountryを実装

+++ Victoria3.GameData/FormableCountry.cs
@@ -12,7 +12,7 @@
         bool IsMajorFormation,
         string? UnificationPlay,
         string? LeadershipPlay,
-        decimal? MaxNumFormationCandidates,
+        int? MaxNumFormationCandidates,
         object? CanBeFormationCandidate,
         object? CanBeUnificationTarget
         )
